Changed the type of pricePerHour for ParkingLot from BigDecimal to Double, so that it is uniform throughout infox widget. Also added animation for clicking on infobox fragment

+++ LotSpot/app/src/main/java/com/utsg/csc301/team21/models/ParkingLot.java
@@ -21,7 +21,7 @@ public class ParkingLot extends AbstractParkingLot implements ClusterItem{
         this.address = address;
         this.lat = lat;
         this.lng = lng;
-        this.pricePerHour = BigDecimal.valueOf( pricePerHour);
+        this.pricePerHour = pricePerHour;
     }
 
     @Override
